Move coding env client to main coding env dir to follow same convention as echo env
