+++ src/envs/coding_env/coding_env_client.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Optional
 from core.http_env_client import HTTPEnvClient
 from core.types import StepResult
 
-from ..models import CodeAction, CodeObservation, CodeState
+from .models import CodeAction, CodeObservation, CodeState
 
 if TYPE_CHECKING:
     from core.containers.runtime import ContainerProvider
